refactor(api): move log logic into separate package
