+++ api/log/doc.go
@@ -1,0 +1,10 @@
+// Copyright (c) 2023 Target Brands, Inc. All rights reserved.
+//
+// Use of this source code is governed by the LICENSE file in this repository.
+
+// Package log provides the log handlers for the Vela API.
+//
+// Usage:
+//
+//	import "github.com/go-vela/server/api/log"
+package log
